requirements.txt
alpaca-trade-api>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0  # Technical analysis library
pytz>=2023.3  # For timezone handling 